Deploy

+++ src/pages/TopicPage.tsx
@@ -1,15 +1,26 @@
 import { useEffect, useState } from "react";
+import TopicChatRoom from "./TopicChatRoom";
 
 const API_URL = import.meta.env.VITE_API_URL;
+interface topic {
+    topicId: string;
+    topicName: string;
+    createdByUser: string;
+}
 
 const TopicPage = () => {
     const [topicName, setTopicName] = useState<string>("");
-    const [topics, setTopics] = useState<any[]>([]);
+    const [topics, setTopics] = useState<topic[]>([]);
+    const [searchQuery, setSearchQuery] = useState<string>("");
+    const [filteredTopics, setFilteredTopics] = useState<topic[]>([]);
+    const [hasJoinedTopic, setHasJoinedTopic] = useState<boolean>(false);
+    const [joinedTopicName, setJoinedTopicName] = useState<string>("");
+    const [topicId, setTopicId] = useState<string>("");
     
     const userId = localStorage.getItem("userId");
     const token = localStorage.getItem("token");
 
-    const createTopic = async (event: any) => {
+    const createTopic = async (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
         if (!token) {
             alert("You need to be logged in to create a topic");
@@ -51,6 +62,7 @@ const TopicPage = () => {
                 }
                 const data = await response.json();
                 setTopics(data);
+                setFilteredTopics(data);
                 console.log(data);
             }  catch (error) {
                     console.error(error);
@@ -61,12 +73,17 @@ const TopicPage = () => {
         useEffect(() => {
             fetchTopics();
     }, []);
-
+        
     const deleteTopic = async (topicId: string) => {
         if (!token) {
             alert("You need to be logged in to delete a topic");
             return;
         }
+        const confirmDelete = window.confirm("Are you sure you want to delete this topic?");
+        if (!confirmDelete) {
+            return;
+        }
+
         try {
             const response = await fetch(`${API_URL}/topic/deleteTopic?topicId=${topicId}&userId2=${userId}`, {
                 method: "DELETE",
@@ -88,37 +105,71 @@ const TopicPage = () => {
         }
     }
 
+    const joinTopicChatRoom = async (topicId: string, topicName: string ) => {
+        console.log("Topic joined");
+        console.log(topicId);
+        setHasJoinedTopic(true);
+        setJoinedTopicName(topicName);
+        setTopicId(topicId);
+    }
+
+    const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
+        setSearchQuery(event.target.value);
+        const filteredTopics = topics.filter((topic) => {
+            return topic.topicName.toLowerCase().includes(event.target.value.toLowerCase());
+        });
+        setFilteredTopics(filteredTopics);
+    }
+
 
 
     return (
         <>
-        <div>
-            <h1>Topic Page</h1>
-            <p>Here you can create a topic, the bot will adjust to the topic and hopefully answer truly!</p>
-            <form onSubmit={createTopic}>
-                <input
-                    className="topicInputField"
-                    type="text"
-                    placeholder="Topic you want to create"
-                    value={topicName}
-                    onChange={(e) => setTopicName(e.target.value)}
-                />
-                <button type="submit">Create topic</button>
-            </form>
-        </div>
-        <div>
-            <h3>Created topics:</h3>
-            {topics.map((topic) => (
-                <div key={topic.topicId}>
-                    <button>{topic.topicName}</button>
-                        {topic.createdByUser === userId && (
-                            <button className="removeTopicBtn" onClick={() => deleteTopic(topic.topicId)}>X</button>
-                        )}
+            {!hasJoinedTopic ? (
+                <>
+                    <div>
+                        <h1>Topic Page</h1>
+                        <p>Here you can create a topic, the bot will adjust to the topic and hopefully answer truly!</p>
+                        <form onSubmit={createTopic}>
+                            <input
+                                className="topicInputField"
+                                type="text"
+                                placeholder="Topic you want to create"
+                                value={topicName}
+                                onChange={(e) => setTopicName(e.target.value)}
+                            />
+                            <button type="submit">Create topic</button>
+                        </form>
+                    </div>
+                    <div>
+                        <h3>Search:</h3>
+                        <input className="searchTopics"
+                            type="text"
+                            placeholder="Search topics"
+                            value={searchQuery}
+                            onChange={handleSearch} />
+                           
+                        <h3>Created topics:</h3> 
+                        <div className="topicGridContainer">
+                        {filteredTopics.map((topic) => (
+                            <div className="topicGrid" key={topic.topicId}>
+                                <button onClick={() => joinTopicChatRoom(topic.topicId, topic.topicName)}>{topic.topicName}</button>
+                                {topic.createdByUser === userId && (
+                                    <button className="removeTopicBtn" onClick={() => deleteTopic(topic.topicId)}>X</button>
+                                )}
+                            </div>
+                        ))}
+                    </div>
                 </div>
-            ))}
-        </div>
+                </>
+            ) : (
+                <div>
+                    <h1>Topic: {joinedTopicName}</h1>
+                    <TopicChatRoom topicId={topicId}/>
+                </div>
+            )}
         </>
     );
-}
+};
 
 export default TopicPage;
